+++ libraries/common/src/main/java/androidx/media3/common/Player.java
@@ -2505,7 +2505,7 @@ public interface Player {
    * @see Listener#onSurfaceSizeChanged(int, int)
    */
   @UnstableApi
-  Size getVideoSurfaceSize();
+  Size getSurfaceSize();
 
   /** Returns the current {@link CueGroup}. */
   CueGroup getCurrentCues();